+++ PHP/deleteUser.php
@@ -1,35 +1,41 @@
 <?php
 // Connect to your database (same as in registration)
-$inData = getRequestInfo();
+if ($_SERVER['REQUEST_METHOD'] === 'DELETE') {
 
-$servername = "localhost";
-$ServerUsername = "phpDealer";
-$ServerPassword = "tTimetocode9!u";
-$dbname = "COP4331";
+    $inData = getRequestInfo();
+    $servername = "localhost";
+    $ServerUsername = "phpDealer";
+    $ServerPassword = "tTimetocode9!u";
+    $dbname = "COP4331";
 
-$conn = new mysqli($servername, $ServerUsername, $ServerPassword, $dbname);
+    $conn = new mysqli($servername, $ServerUsername, $ServerPassword, $dbname);
 
-// Check connection
-if ($conn->connect_error) {
-    returnWithError($conn->connect_error);
+    // Check connection
+    if ($conn->connect_error) {
+        returnWithError($conn->connect_error);
+    }
+    else
+    {
+        //BEFORE RUNNING THIS, MAKE SURE THE ID PASSED FROM JSON IS NOT CONFLATED WITH USERID
+        $stmtDelete = $conn->prepare("DELETE FROM Contacts WHERE ID = ?");
+        $stmtDelete->bind_param("i", $inData['id']);
+        if (!$stmtDelete->execute())
+        {
+            returnWithError("Failed to Delete User");
+        }
+        //BEFORE RUNNING THIS, MAKE SURE THE ID PASSED FROM JSON IS THE USERID
+        $stmtDelete2 = $conn->prepare("DELETE FROM Users WHERE ID = ?");
+        $stmtDelete2->bind_param("i", $inData['userid']);
+        if (!$stmtDelete->execute())
+        {
+            returnWithError("Failed to Delete User");
+        }
+        $stmtDelete2->close();
+    }
 }
 else
 {
-    //BEFORE RUNNING THIS, MAKE SURE THE ID PASSED FROM JSON IS NOT CONFLATED WITH USERID
-    $stmtDelete = $conn->prepare("DELETE FROM Contacts WHERE ID = ?");
-    $stmtDelete->bind_param("i", $inData['id']);
-    if (!$stmtDelete->execute())
-    {
-        returnWithError("Failed to Delete User");
-    }
-    //BEFORE RUNNING THIS, MAKE SURE THE ID PASSED FROM JSON IS THE USERID
-    $stmtDelete2 = $conn->prepare("DELETE FROM Users WHERE ID = ?");
-    $stmtDelete2->bind_param("i", $inData['userid']);
-    if (!$stmtDelete->execute())
-    {
-        returnWithError("Failed to Delete User");
-    }
-    $stmtDelete2->close();
+    returnWithError("Invalid Request Method");
 }
 //Helper Functions
 function getRequestInfo()
